Refactor test classes to use new abstract base class so that tests in SandboxTransformerTest do not run twice
